# api/handler/product_handler.go
package handler

import (
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
	"gocashier.db/internal/models"
	"gocashier.db/internal/services"
)

type productHandler struct {
	productService services.ProductService
}

func NewProductHandler(productService services.ProductService) *productHandler {
	return &productHandler{
		productService: productService,
	}

}


// Create godoc
// @Summary Create product
// @Tags products
// @Accept json
// @Produce json
// @Param product body models.Product true "Product Data"
// @Success 200 {object} models.Product
// @Router /products [post]
func (c *productHandler) Create(h *gin.Context) {
	var product models.Product
	err := h.ShouldBindJSON(&product)
	if err != nil {
		h.JSON(http.StatusBadRequest, gin.H{
			"error": "invalid request payload",
		})
		return
	}

	err = c.productService.Create(&product)
	if err != nil {
		h.JSON(http.StatusBadRequest, gin.H{
			"error": err.Error(),
		})
		return
	}

	h.JSON(http.StatusOK, gin.H{
		"success": true,
		"data":    product,
	})

}


// GetAll godoc
// @Summary Get all products
// @Tags products
// @Produce json
// @Success 200 {array} models.Product
// @Router /products [get]
func (c *productHandler) GetAll(h *gin.Context) {

	data, err := c.productService.GetAll()
	if err != nil {
		h.JSON(http.StatusBadRequest, gin.H{
			"error": err.Error(),
		})
		return
	}

	h.JSON(http.StatusOK, gin.H{
		"succsess": true,
		"data":     data,
	})

}


// UpdateById godoc
// @Summary Update product by ID
// @Tags products
// @Accept json
// @Produce json
// @Param ID path int true "Product ID"
// @Param product body models.Product true "Product Data"
// @Success 200 {object} models.Product
// @Router /products/{ID} [put]
func (c *productHandler) UpdateById(h *gin.Context) {
	id, err := strconv.Atoi(h.Param("ID"))
	if err != nil {
		h.JSON(http.StatusBadRequest, gin.H{"error": "Invalid ID"})
		return
	}

	var product models.Product
	if err := h.ShouldBindJSON(&product); err != nil {
		h.JSON(http.StatusBadRequest, gin.H{
			"error": "invalid request payload",
		})
		return

	}

	if err = c.productService.UpdateById(id, &product); err != nil {
		h.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	h.JSON(http.StatusOK, gin.H{
		"succsess": true,
		"data":     product,
	})
}


// DeleteById godoc
// @Summary Delete product by ID
// @Tags products
// @Produce json
// @Param ID path int true "Product ID"
// @Success 200 {string} string "Success delete data"
// @Router /products/{ID} [delete]
func (c *productHandler) DeleteById(h *gin.Context) {
	id, err := strconv.Atoi(h.Param("ID"))
	if err != nil {
		h.JSON(http.StatusBadRequest, gin.H{"error": "Invalid ID"})
		return
	}

	if err = c.productService.DeleteById(id); err != nil {
		h.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	h.JSON(http.StatusOK, gin.H{
		"succsess": true,
		"message":  "Success delete data",
	})
}


// GetById godoc
// @Summary Get product by ID
// @Tags products
// @Produce json
// @Param ID path int true "Product ID"
// @Success 200 {object} models.Product
// @Router /products/{ID} [get]
func (c *productHandler) GetById(h *gin.Context) {
	id, err := strconv.Atoi(h.Param("ID"))
	if err != nil {
		h.JSON(http.StatusBadRequest, gin.H{"error": "Invalid ID"})
		return
	}
	data, err := c.productService.GetById(id)
	if err != nil {
		h.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	h.JSON(http.StatusOK, gin.H{
		"succsess": true,
		"data":     data,
	})
}



// GetDetailProductById godoc
// @Summary Get detail product by ID
// @Tags products
// @Produce json
// @Param ID path int true "Product ID"
// @Success 200 {object} models.Product
// @Router /products/{ID}/detail [get]
func (c *productHandler) GetDetailProductById(h *gin.Context) {
	id, err := strconv.Atoi(h.Param("ID"))
	if err != nil {
		h.JSON(http.StatusBadRequest, gin.H{"error": "Invalid ID"})
		return
	}

	data, err := c.productService.GetDetailProductById(id)
	if err != nil {
		h.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	h.JSON(http.StatusOK, gin.H{
		"succsess": true,
		"data":     data,
	})

}